backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.4.0
google-cloud-vision==3.7.2
sentence-transformers==3.0.1
numpy==2.1.2
scipy==1.13.1
Pillow==10.4.0
aiofiles==23.2.1

